Adjustments made to OpenGLView and OpenGLItem
OpenGLView

Dimensions for the cubes have been moved to separate floats and referenced in the constructors for the walls. A cube has also been added for test purposes

OpenGLItem

Added a virtual setPosition slot which has been redefined in OpenGLItemCuboid. Moved the position and dimension vectors as private elements for the class, moved the vertices construction to a sepeate function which references the position and dimensions.